Server. Basic refactoring and splitting.
