+++ backend/firebase.js
@@ -13,4 +13,3 @@ const config = {
 firebase.initializeApp(config);
 
 module.exports = firebase;
-
